// frontend/src/app/api/model/pageMetadata.ts
/**
 * My API
 * OpenAPI documentation for My API
 *
 * The version of the OpenAPI document: 59acdd34ea678de65cbe0a25f66df8cefdf5050b
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PageMetadata { 
    size: number;
    number: number;
    totalElements: number;
    totalPages: number;
}